feat(invoice-module): create tranched stream with timestamps, improve checks and types

+++ src/modules/invoice-module/libraries/Types.sol
@@ -20,7 +20,7 @@ library Types {
         // slot 0
         Method method;
         Recurrence recurrence;
-        uint24 paymentsLeft;
+        uint40 paymentsLeft;
         address asset;
         // slot 1
         uint128 amount;
